+++ dom/base/ScreenOrientation.cpp
@@ -18,6 +18,7 @@
 #include "mozilla/dom/ContentChild.h"
 #include "mozilla/dom/Event.h"
 #include "mozilla/dom/Promise.h"
+#include "mozilla/StaticPrefs_browser.h"
 #include "nsContentUtils.h"
 
 using namespace mozilla;
@@ -158,8 +159,11 @@ bool ScreenOrientation::LockOrientationTask::OrientationLockContains(
 
 NS_IMETHODIMP
 ScreenOrientation::LockOrientationTask::Run() {
-  // Step to lock the orientation as defined in the spec.
+  if (!mPromise) {
+    return NS_OK;
+  }
 
+  // Step to lock the orientation as defined in the spec.
   if (mDocument->GetOrientationPendingPromise() != mPromise) {
     // The document's pending promise is not associated with this task
     // to lock orientation. There has since been another request to
@@ -182,18 +186,28 @@ ScreenOrientation::LockOrientationTask::Run() {
     return NS_OK;
   }
 
-  ErrorResult rv;
-  bool result = mScreenOrientation->LockDeviceOrientation(mOrientationLock,
-                                                          mIsFullscreen, rv);
-  if (NS_WARN_IF(rv.Failed())) {
-    return rv.StealNSResult();
-  }
+  RefPtr<MozPromise<bool, bool, false>> lockOrientationPromise =
+      mScreenOrientation->LockDeviceOrientation(mOrientationLock,
+                                                mIsFullscreen);
 
-  if (NS_WARN_IF(!result)) {
+  if (NS_WARN_IF(!lockOrientationPromise)) {
     mPromise->MaybeReject(NS_ERROR_UNEXPECTED);
     mDocument->ClearOrientationPendingPromise();
     return NS_OK;
   }
+
+  lockOrientationPromise->Then(
+      GetCurrentSerialEventTarget(), __func__,
+      [self = RefPtr{this}](
+          const mozilla::MozPromise<bool, bool, false>::ResolveOrRejectValue&
+              aValue) {
+        if (aValue.IsResolve()) {
+          return NS_OK;
+        }
+        self->mPromise->MaybeReject(NS_ERROR_DOM_ABORT_ERR);
+        self->mDocument->ClearOrientationPendingPromise();
+        return NS_OK;
+      });
 
   BrowsingContext* bc = mDocument->GetBrowsingContext();
   if (OrientationLockContains(bc->GetCurrentOrientationType()) ||
@@ -304,6 +318,12 @@ already_AddRefed<Promise> ScreenOrientation::LockInternal(
   p->MaybeReject(NS_ERROR_DOM_NOT_SUPPORTED_ERR);
   return p.forget();
 #else
+  // Bypass locking screen orientation if preference is false
+  if (!StaticPrefs::dom_screenorientation_allow_lock()) {
+    p->MaybeReject(NS_ERROR_DOM_NOT_SUPPORTED_ERR);
+    return p.forget();
+  }
+
   LockPermission perm = GetLockOrientationPermission(true);
   if (perm == LOCK_DENIED) {
     p->MaybeReject(NS_ERROR_DOM_SECURITY_ERR);
@@ -342,11 +362,10 @@ already_AddRefed<Promise> ScreenOrientation::LockInternal(
 #endif
 }
 
-bool ScreenOrientation::LockDeviceOrientation(
-    hal::ScreenOrientation aOrientation, bool aIsFullscreen, ErrorResult& aRv) {
+RefPtr<MozPromise<bool, bool, false>> ScreenOrientation::LockDeviceOrientation(
+    hal::ScreenOrientation aOrientation, bool aIsFullscreen) {
   if (!GetOwner()) {
-    aRv.Throw(NS_ERROR_UNEXPECTED);
-    return false;
+    return MozPromise<bool, bool, false>::CreateAndReject(false, __func__);
   }
 
   nsCOMPtr<EventTarget> target = GetOwner()->GetDoc();
@@ -355,11 +374,7 @@ bool ScreenOrientation::LockDeviceOrientation(
   // This needs to be done before LockScreenOrientation call to make sure
   // the locking can be unlocked.
   if (aIsFullscreen && !target) {
-    return false;
-  }
-
-  if (NS_WARN_IF(!hal::LockScreenOrientation(aOrientation))) {
-    return false;
+    return MozPromise<bool, bool, false>::CreateAndReject(false, __func__);
   }
 
   // We are fullscreen and lock has been accepted.
@@ -368,15 +383,21 @@ bool ScreenOrientation::LockDeviceOrientation(
       mFullscreenListener = new FullscreenEventListener();
     }
 
-    aRv = target->AddSystemEventListener(u"fullscreenchange"_ns,
-                                         mFullscreenListener,
-                                         /* useCapture = */ true);
-    if (NS_WARN_IF(aRv.Failed())) {
-      return false;
+    nsresult rv = target->AddSystemEventListener(u"fullscreenchange"_ns,
+                                                 mFullscreenListener,
+                                                 /* aUseCapture = */ true);
+    if (NS_WARN_IF(NS_FAILED(rv))) {
+      return MozPromise<bool, bool, false>::CreateAndReject(false, __func__);
     }
   }
 
-  return true;
+  RefPtr<MozPromise<bool, bool, false>> halPromise =
+      hal::LockScreenOrientation(aOrientation);
+  if (halPromise == nullptr) {
+    return MozPromise<bool, bool, false>::CreateAndReject(false, __func__);
+  }
+
+  return halPromise;
 }
 
 void ScreenOrientation::Unlock(ErrorResult& aRv) {
@@ -521,7 +542,7 @@ void ScreenOrientation::Notify(const hal::ScreenConfiguration& aConfiguration) {
   if (doc->Hidden() && !mVisibleListener) {
     mVisibleListener = new VisibleEventListener();
     rv = doc->AddSystemEventListener(u"visibilitychange"_ns, mVisibleListener,
-                                     /* useCapture = */ true);
+                                     /* aUseCapture = */ true);
     NS_WARNING_ASSERTION(NS_SUCCEEDED(rv), "AddSystemEventListener failed");
     return;
   }
@@ -541,8 +562,14 @@ void ScreenOrientation::UpdateActiveOrientationLock(
   if (aOrientation == hal::eScreenOrientation_None) {
     hal::UnlockScreenOrientation();
   } else {
-    DebugOnly<bool> ok = hal::LockScreenOrientation(aOrientation);
-    NS_WARNING_ASSERTION(ok, "hal::LockScreenOrientation failed");
+    hal::LockScreenOrientation(aOrientation)
+        ->Then(
+            GetMainThreadSerialEventTarget(), __func__,
+            [](const mozilla::MozPromise<bool, bool,
+                                         false>::ResolveOrRejectValue& aValue) {
+              NS_WARNING_ASSERTION(aValue.IsResolve(),
+                                   "hal::LockScreenOrientation failed");
+            });
   }
 }