+++ mobile/android/geckoview/src/main/java/org/mozilla/geckoview/doc-files/CHANGELOG.md
@@ -26,11 +26,13 @@ exclude: true
   HTML content as plain text.
 - Removed deprecated Content Blocking APIs.
   ([bug 1743706]({{bugzilla}}1743706)).
+- Added [`OrientationController`][96.5] to allow GeckoView to handle orientation locking.
 
 [96.1]: {{javadoc_uri}}/Autocomplete.StorageDelegate.html#onLoginFetch--
 [96.2]: {{javadoc_uri}}/GeckoResult.html#finally_-java.lang.Runnable-
 [96.3]: {{javadoc_uri}}/WebExtension.InstallException.ErrorCodes.html#ERROR_INVALID_DOMAIN-
 [96.4]: {{javadoc_uri}}/GeckoSession.SelectionActionDelegate.Selection.html#pasteAsPlainText--
+[96.5]: {{javadoc_uri}}/OrientationController.html
 
 ## v95
 - Added [`GeckoSession.ContentDelegate.onPointerIconChange()`][95.1] to notify
@@ -1091,4 +1093,4 @@ to allow adding gecko profiler markers.
 [65.24]: {{javadoc_uri}}/CrashReporter.html#sendCrashReport-android.content.Context-android.os.Bundle-java.lang.String-
 [65.25]: {{javadoc_uri}}/GeckoResult.html
 
-[api-version]: f8a31bdc9d6debf2e34c1a231d13bfaad76a08aa
+[api-version]: 2e5d064d5675b5e38eeb25595a11906ed8d43e1a